test: split connector and PayPal delivery tests
